+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.0.1 - 2025-12-27
+
+* migrate from kivitendo-erp to Kiwifree-ERP
+* add mising dependency
+* rename project and rest-api
+* add task-server service
+
 ## 3.9.2 - 2025-07-18
 
 * use latest kivitendo 3.9.2